Split off rpmDigest*() APIs from rpmpgp.h to a new rpmcrypto.h header
These APIs never belonged to rpmpgp.h anyway, it was only used for
being the least-worst fit within rpm. As if it was so hard to just
add a new header...
