+++ rpmio/Makefile.am
@@ -23,7 +23,7 @@ librpmio_la_SOURCES = \
 	rpmio_internal.h rpmhook.h rpmvercmp.c rpmver.c \
 	rpmstring.c rpmfileutil.c rpmglob.c \
 	rpmkeyring.c rpmstrpool.c rpmmacro_internal.h \
-	rpmlua.c rpmlua.h
+	rpmlua.c rpmlua.h rpmcrypto.h
 
 if WITH_OPENSSL
 librpmio_la_SOURCES += digest_openssl.c
